// src/app/shared/models/periodo.ts
export interface Periodo {

    nombreCorto?: string;
	idEmpresa?: number;
    idPeriodo?: number;
    idFrecuencia?: number;
    dePeriodo?: string;
    deAnioPeriodo?: string;
    feIni?: string;
    feFin?: string;
    anio?: number;
    inBaja?: string;
    idUsuaCrea?: string;
    deUsuaCreaIp?: string;
    feUsuaCrea?: string;
    idUsuaModi?: string;
    deUsuaModiIp?: string;
    feUsuaModi?: string;
    // usuarioModificacion?: string;
    // fechaModificacion?: string;
    // ipModificacion?: string;
}